Extend default dentry-cache-timeout ival and expose cli knob
[ci skip]

The goal here is to reduce the number of lookup() operations receibed from VFS, and in consequence, bring down the number of 'nsenter' episodes.
I'm extending the existing default timer from 1min to 5min. I'm also exposing a CLI knob to allow end-users to set this value as they wish.