Fix mount-syscall issue preventing relative-paths to be properly verified
In the current code, path attributes of all incoming mount instructions are converted to 'absolute' representation.
In consequence, our verification logic can't really asses the suitability of paths passed in 'relative' form. This
can potentially represent a security issue as we could end up granting access to users that have not the proper/enough
credentials.

The fix is to defer this relative-to-absolute conversion till the paths verification logic has concluded.