Modify containerDB and handler-processing logic to index cntrs by user-ns-inode
Currently, all container-state in sysbox-fs is indexed by pid-ns-inode. This
approach has some constrains when dealing with nested container scenarios, as
we are unable to iterate the namespace hierarchy upwards (from inner to outer
containers).

This patch modifies containerDB to utilize user-ns-inode as the new index.
The assumption we are making here is that within a sysbox sys-container there
won't ever be a requirement to have more than one child user-namespace. In
consequence, there's no need to iterate the namespace hierarchy as all inner
containers will necessarilly be either: 1) part of the sys-container user-ns,
or 2) they are part of the sys-container child user-namespace.