Fix bug in common handler when entering sys container namespaces.
The common handler must enter the namespaces of the process requesting
an access, rather than the namespaces of the sys container's init
process.

This ensures that if the process performing the access is inside
an inner container (or with unshared namespaces), the common
handler will deal with the access correctly.

For example, if a sys container is launched and an inner container is
created, sysbox-fs backs the /proc/sys directory in both the sys
container and inner container. If a process inside the sys container
accesses a namespaced resource (e.g., /proc/sys/net/ipv4/ip_forward),
the common handler should enter the sys container namespaces and
perform the access. But if a process inside the inner container the
reads that same resource, the common handler must enter the namespaces
of the inner container, not those of the sys container, since the
inner container will have its own version of those resources.