Remove sysbox-fs process camouflaging.
There is no current need for the sysbox-fs nsenter process to acquire
the personality of another process. As a result, the process
camouflaging functionality is not required.

The camouflaging functionality was added earlier, when it was thought
that the sysbox-fs nsenter process needed to have the same uid, gid,
and caps as the process that initiated the access to /proc/sys from
within a sys container, in order to ensure process permissions where
checked correctly during the access. But since we've realized that
this is not necessary, as the sysbox-fs FUSE lib supports a config
that ensures that the Linux VFS performs those checks before the
access arrives at sysbox-fs.