Fix sysbox-fs lookup attributes.
This change fixes a bug in sysbox-fs, in which the FUSE Lookup() function
was incorrectly collecting the uid & gid attributes of procfs files.

This bug was causing an inconsistency with the uid & gid attributes
that sysbox-fs was passing in FUSE responses: for Lookup(), the
incorrect attributes were being passed, for Getattr() the correct
attributes were being passed.