Remove sysbox-fs left-over child process reaper.
In a prior change, we added a go-routine in sysbox-fs to perform
reaping of left-over child processes that occassionally show
up when sysbox-fs dispatches the nsenter agent as part of
emulating procfs or mount syscalls inside a sys container.

The reaper worked by simply periodically reaping any child zombie
processes.

But it turns out this simple approach is not good enough, as it's
possible for the reaping to conflict with normal reaping done
by the nsenter dispatcher, causing the dispatcher to generate
errors that can hang the system container.

This change removes the child reaper, and instead improves child
reaping in the nsenter dispatcher. With these improvements I don't
see left-over child zombie processes after several test runs.