Fix for issue preventing emulated nodes to be properly lookup
Problem impacts ReadDirAll() operation, specifically on folders that contain emulated resources.
The existing logic would differentiate between ADDED and SUBSTITUTED nodes to address the
difference between nodes that must be explicitly exposed (to be added), and those that must be simply
replaced for various reasons (refer to code for details).

The problem with the above approach is that it requires a static labeling of FS nodes. This poses a
challenge when dealing with various Linux distros or kernel-releases, as nodes 'exposure' varies
accross these scenarios.

The fix is to add some extra logic to ReadDirAll() operation to detect overlapping nodes and eliminate
them if needed. This approach also allows the elimination of one of the node states previously utilized
(ADDITION).