Trap and emulate chown to /sys inside a sys container.
This commit adds chown syscall trapping & handling code to sysbox-fs.

The reason we trap chown (as well as fchown and fchownat) is to prevent
chown to "/sys" inside a sys container from failing.

The reason chown to "/sys" inside a sys container would fail without this code is
that "/sys" is owned by the host's true root, so it shows up as "nobody:nogroup"
inside the sys container and thus its ownership can't be changed from within the
container. Some apps running inside the container (e.g,. RedHat's RPM package
manager) want to chown "/sys" to root:root, causing the apps to get an EPERM and
fail.

As a work-around, Sysbox ignores chown to "/sys" inside the sys container (or in
any inner containers). All other chown operations are handled normally by the
kernel.

Note that unlike procfs, we can't chown "/sys" to "root:root" inside the containers
because sysfs ownership if global across all mountpoints of the host. That is,
when ownership is changed inside a container, the change is seen in all sysfs
mounts on the host.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>