Relax seccomp-tracer's processing logic to iterate upon syscall error detection
There is room in the existing code for an spurious error in the seccomp-tracer causing a halt in the syscall-trapping processing.

This PR tackles this issue by relaxing the seccomp-notification error-handling logic. If a legit error were to be found in NotifReceive()/NotifResponse(), this one will be catched by the routine in charged of verifying seccomp-fd's health, and only then the seccomp-tracer logic will stop processing syscalls for the affected sys-container.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>