Improve sysbox-fs handlers to deal with concurrent accesses to emulated resources.
It's possible that within a sys container, multiple processes may be accessing
the same sysbox-fs backed resources under procfs and sysfs (e.g., /proc/swaps,
/proc/sys/net/netfilter/nf_conntrack_max, etc). To ensure consistency across
these accesses, sysbox-fs needs to implement per-resource, per-container
locking. Such locking partially existed, but was not sufficient. This commit
fixes this.

In addition, some of the sysbox-fs handlers that emulate procfs and sysfs
resources must sometimes write to the kernel's real procfs and sysfs.  In this
case, an additional per-resource locking across all containers is required, to
ensure sysbox-fs writes consistent values to the real procfs or sysfs.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>