Fix deadlock in sysbox-fs seccomp tracer.
This change fixes a deadlock that occurs in the sysbox-fs
seccomp tracer under load pressure (i.e., when many
system containers are launched quickly).

The deadlock occurred in the interaction between the
seccomp tracer's sessionsMonitor and the pidMonitor:
the sessionsMonitor was not able to add events
to the pidMonitor, because the pidMonitor was blocked
waiting for the sessionsMonitor to read events; and
the sessionsMonitor would not read events until it
was able to add events first.

This in turn was blocking sysbox-fs from acknowledging
reception of seccomp fds from sysbox-runc, which in
turn was blocking sysbox-runc and thus the container
engine above it (e.g., Docker/containerd).

The fix is to decouple the pid monitor add events vs. read events
in the sysbox-fs tracer. The add events are now done synchronously
when receiving new seccomp fds from sysbox-runc. The read-events
are done asynchronously in the sessionsMonitor goroutine.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>