Improve the cap_last_cap handler to reduce number of open file descriptors.
Prior to this change, the sysbox-fs handler for /proc/sys/kernel/cap_last_cap
would open a file descriptor whenever a process in a container opened this
resource. And it would keep it open until the process in the container closed
the resource. Given that cap_last_cap is one of the most (if not the most)
accessed sysctl by processes inside containers, this led to sysbox-fs having a
very large number of file descriptors opened at a given time.

This change improves on this, by having sysbox-fs only open the cap_last_cap
handler when it's value is read, and closes it immediately after.  This
significantly reduces the number of open fds that sysbox-fs has at a any given
time.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>