sysbox-fs fuse: return the requester's user-ns root uid & gid in the lookup attributes.
When the sysbox-fs FUSE server receives a request from a process inside a sys container
to access procfs or sysfs, the file owner it returns should correspond to the
root user in the user-ns associated with the process doing the request.

Prior to this, the code was returning the file owner corresponding to the root user
in the sys container of the process doing the request, which is an approximation
but not quite correct (since inside the sys container there could be child user-ns).

As part of this change, I refactored a bit and created a new method around the
process class that allows a process to find out the uid & gid map for the root user
in it's user-ns.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>