Make containers that share the same net-ns share fuse-server state.
In order to support K8s pods, we should use the network-ns as the indicators
that containers belong to the same pod.

Prior to this change, sysbox-fs was using the user-ns to determine which
containers belong to the same pod. But this is incorrect, as the K8s KEP-127
proposes a "cluster" mode in which *all* pods in the host would share the same
user-ns. In fact this is the initial proposed mode of operation for rootless
pods in K8s, until "pod" mode is implemented (where each pod is in a dedicated
user-ns for stronger isolation).

With this commit, sysbox-fs uses both the network-ns, rather than the user-ns,
to determine if multiple containers are in the same pod (i.e., if they should
share sysbox-fs state).

Note that sharing a net-ns implies sharing a user-ns, since the net-ns
is "owned" by a user-ns (see user_namespaces(7)).

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>