Relax error check during container unregistration.
This change relaxes an error check during container unregistration. It's needed
to deal properly with cases where a container pre-registers and then unregisters
due to an error that occurs when starting the container (e.g., on sysbox-runc).

Without this change, such a situation results in sysbox-fs leaving stale
state for the failed container.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>