Fix issue preventing virtual nodes from being assigned a unique / non-zero inode value
As part of this change i'm bringing back the Attr() method to obtain the details of every FUSE-emulated node.

Upon arrival of an Attr() or GetAttr() request from VFS, Bazil FUSE-lib interacts with the app implementing the FUSE-server functionality to obtain the FS attributes of that given node. Now, regardless of the incoming syscall (Attr or Getattr), Bazil invokes either the Attr() or the Getattr() callback methods to fetch the required information from the app implementing the FUSE-server functionality (i.e. sysbox-fs in this case). Bazil picks whichever callback is implemented by the FUSE-server as long as this one satisfies the fuse.Node interface.

Back in the days when we started the sysbox-fs implementation, i found the Attr() approach to be more intuitive than the Getattr() one, but then decided to switch to GetAttr() callback given that its signature provided more information about the incoming syscall generated by the user. Unfortunately, i just discovered a non-trivial bug in Bazil implementation that prevents it from being able to generate unique inode-values for sysbox-fs nodes that are not backed up by real procfs/sysfs node entries; IOW, Bazil can't generate unique inode values whenever the received node is lacking one (inode == 0).

A fix for this issue in Bazil would take some time as it would require some important implementation changes. Luckily, i realized that the code-path offered by the Attr() callback is not affected by this bug, so i'm switching sysbox-fs implementation back to the Attr() approach, which won't have any impact or our existing implementation as we are not relying anymore on that "extra" information offered by GetAttr().

Signed-off-by: Rodny Molina <rmolina@nestybox.com>