Improve resolution of "/proc/self/*" paths in syscall interception code.
This change improves the resolution of "/proc/self/*" paths in syscalls
intercepted by Sysbox. The resolution means that the /proc/self/* is turned into
/proc/<pid>/*, where <pid> is corresponds to the process that called the syscall
(i.e., the target process)

Prior to this change, this path resolution was after the sysbox-fs nsenter agent
had entered the namespaces of the target process. But it turn out this fails in
scenarios where /proc is not mounted within the target's process mount ns, or
more subtly in scenarios where it's mounted but not remounted within the target
process's pid namespace. In both cases, the "/proc/self" resolution fails since
the nsenter agent won't find the pid of the target process in procfs (or will
find an incorrect one).

The latter (more subtle) scenario was discovered in Sysbox issue #291, where a
recent change in the OCI runc (release v1.0.0-rc95) caused failures when running
Docker inside a Sysbox container.  Specifically, the inner OCI runc changed the
way it mounts /proc in the inner container, from "mount -t proc proc
/path/to/inner/container" to "mount -t proc proc /proc/self/fd/7", where
"/proc/self/fd/7" is a symlink to "/path/to/inner/container". This requires
Sysbox to resolve "/proc/self" to "/proc/pid", but this failed when the
sysbox-fs nsenter agent tried doing this because the proc in the inner
container was not yet in the inner container pid ns (i.e., rather it was the
proc inherited from the parent runc process), so the resolution failed.

This commit corrects this, by moving the resolution of /proc/self from the
sysbox-fs nsenter agent to the sysbox-fs main process. This way the resolution
occurs in the context of sysbox-fs, where proc is always mounted and sysbox will
always resolve the target process's /proc/self to /proc/<pid> correctly, thereby
allowing it to resolve /proc/self/* symlinks properly.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>