Modify existing 'per-handler' collision-detection logic to make it a 'per-resource' one
This change fixes an issue with the existing logic in charge of preventing collisions when writing to system-wide resources. Currently, this logic is 'per-handler' and not 'per-resource', and as such, may unnecessarily defer I/O for resources for which no contention exists.

The fix is to define locks for every shared resource and acquire these locks when required.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>