Move and rename 'procSysCommon' handler as 'passthrough' one for efficiency and consistency purposes
This change adds more clarity to the purpose of the 'pass-through' handler, which was tied up to the '/proc/sys' subtree. Not only was this conceptually inacurate (i.e., a pass-through handler should be able to serve any subtree), but also contributed to unnecessary cycles being wasted when serving I/O operations. By holding a pointer to the pass-through handler within the handler-service object, we are easing the way in which regular handlers reference the pass-through one.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>