Add config option to release seccomp-notify fd handles on container exit.
This commit adds a config option to sysbox-fs called "seccomp-fd-release". It
solves the problem of syscall interception not working when a process execs into
a container, forks children, and then dies, causing sysbox to close the syscall
interception handle (seccomp-notify fd) and thus breaking syscall interception
for the children processes.

When set to "cont-exit" (container exit), sysbox-fs will not close the
syscall interception handle when the associated process dies but instead will
keep it open until the associated container stops, solving the exec problem. By
default, this option is set to "proc-exit" (process exit), meaning that
sysbox-fs closes the syscall interception handle as soon as the associated
process dies (i.e., no change in the behavior prior to this change).

While setting this new config option to "cont-exit" fixes the problem described
above, the drawback that this causes kernel resources associated with that
handle to be kept around even after the associated process has died. For long
living containers where lot's of processes exec into the container, this could
be a problem so it's not recommended.  However, for short lived containers where
processes exec into the container, this should be fine.

Note that this config option is expected to be temporary because starting in
kernel >= 5.8, there is a kernel mechanism to help sysbox determine when *all*
processes associated with a given seccomp notify fd have died, thus solving the
problem described above. As a result, this option is expected to be deprecated
(or become useless) in the future.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>