Fix a bug related to the seccomp-fd-release config option.
A prior commit added the "seccomp-fd-release" config option to sysbox-fs, which
allows sysbox-fs to support syscall interception of processes that "exec" into
the container, fork, and then die.

However there was a bug: the sysbox-fs seccomp tracer was removing the dying
process from the seccompSessionMap, thereby preventing rearming the seccomp
notification after one invocation. This change fixes this by causing the seccomp
tracer to keep the pid in the seccompSessionMap until the container dies.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>