Fix bug preventing ARM platforms from being able to read RO procfs nodes
Discovered this one while running some preliminary tests on ARM platforms. It turns out that 'cat' application when running over ARM, seems to be calling open() with "O_READ | O_NOFOLLOW" flags, which resulted in Sysbox returning an EACCESS back to the user (which can be easily explained by looking at the existing code).

This same bug can be potentially seen in AMD64 arch too, but we haven't run into this till we started to play with ARM.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>