Fix bug in sysbox-fs emulation of l*xattr() syscalls.
The l*xattr() syscalls take a path to a file, and if that file
is a symlink, they must not follow it.

The pathAccess() method of the process object was not honoring
this and always following the symlink. This caused failures
when the symlink was dangling.

An example of such a failure was discovered in Sysbox issue #415,
where running K8s.io KinD inside a sysbox container caused
the container to call to execute llistxattr(), which was then
trapped by Sysbox but not emulated correctly. As a result,
KinD failed with the following message:

"failed to list xattrs on /var/lib/docker/overlay2/9e0c206a7cf7a33d50da94fe146f7ed1a13604a7f3f1d448b7124d2540a461d2/merged/var/lock: no such file or directory"

This commit fixes this and adds a unit test to verify the fix.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>