Adjust seccomp-bfd main loop to avoid exiting upon notif() error arrival
This one gave me a few headaches while trying to add support for buildx and improve the overall sysbox-fs performance.

Problem would manifest sporadically when attempting to build docker images within a sysbox-powered buildkit container:

    ```
    vagrant@ubuntu-focal-vm-2:~/dockerfile-examples$ docker --debug buildx build --load -t nestybox/delete-me:latest --platform linux/amd64 .
    [+] Building 0.0s (0/0)
    [+] Building 0.0s (0/0)
    [+] Building 0.0s (0/0)
    [+] Building 14.3s (5/11)
    [+] Building 14.3s (6/11)
     => [internal] load build definition from Dockerfile
     ...
     => ERROR [2/7] WORKDIR /usr/node                                                                                                                                                                                                         0.0s
    ------
     > [2/7] WORKDIR /usr/node:
    ------
    Dockerfile:8
    --------------------
       6 |     ENV workdirectory /usr/node
       7 |
       8 | >>> WORKDIR $workdirectory
       9 |     WORKDIR app
      10 |
    --------------------
    error: failed to solve: failed to compute cache key: failed to unmount /tmp/containerd-mount077965446: function not implemented: mount callback failed on /tmp/containerd-mount077965446: lchown /tmp/containerd-mount077965446/usr/local/lib/node_modules/npm/node_modules/pacote/node_modules/make-fetch-happen/node_modules/https-proxy-agent/node_modules/agent-base/node_modules/es6-promisify/node_modules/es6-promise/LICENSE: function not implemented
    1 v0.9.3 buildkitd
    ```

The error above is a direct consequence of the seccomp tracert logic exiting its main event-loop for the seccomp-fd associated with the buildx's control session. The error is not consistently reproduced coz it requires that kernel considers this seccomp-fd to be temporarily unsuitable for operation, and as such returns ENOENT during ioctl(seccomp_notif_receive()) execution. In this scenario, we shouldn't bail out, but simply retry the notif_receive() call. There are other mechanisms in place to detect and halt the processing of permanently-invalid seccomp-fd.

Signed-off-by: Rodny Molina <rmolina@nestybox.com>