Lower sysbox-fs log message from error to debug.
The error log in the fuse ReadDirAll() when the handler fails was a bit strong
since it can occur under normal conditions when a handler fails to nsenter into
the container due to the container dying at the time the access occurs.

This was causing some performance tests to fail when they check the sysbox-fs
log for errors.

Lower it to debug, same as we do in fuse Create() function.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>