Refactor and simplify procfs and sysfs handlers.
The "/proc" handler was removed since we don't emulate or mount sysbox-fs at
that level; instead, the "/proc/uptime" and "/proc/swaps" handlers were created
and added to the list of handlers.

In addition, the "/proc/sys" handler was simplified a bit to deal with
all accesses to /proc/sys.

Also, the "/sys" handler was merged into the "/sys/kernel" handler for the
same reason.

Finally, the methods of the refactored handlers were adjusted appropriately.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>