Fix implementation of handler for /sys/module/nf_conntrack/parameters.
Prior to this change, the handler was emulating the
/sys/module/nf_conntrack/parameters/hashsize file only, even though the
handler's path was set to emulate at /sys/module/nf_conntrack/parameters.

Though this is incorrect, the reason it worked before is because sysbox-runc was
mounting sysbox-fs at /sys/module/nf_conntrack/parameters/hashsize.

With this change, we fix the handler to emulate at /sys/module/nf_conntrack/parameters.
A change will be made in sysbox-runc to mount it at that same directory.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>

Fix to nf_conntrack handler.