Sort procfs and sysfs mountpoints hierarchically.
sysbox-fs mountpoints over procfs and sysfs should ideally be sorted
hierarchically in case a latter mount depends on the earliers one (currently not
the case).

For example, sysbox-fs currently mounts itself over the sysbox container's sysfs
at "/sys/kernel". If we ever have a submount of sysbox-fs under "/sys/kernel/*"
or if we ever depend on the "/sys/kernel" mount to be present before performing
other operations, it's cleaner to mount that one before the submounts.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>