Fix for Sysbox v0.6.1 regression on emulation of /sys/devices/virtual/*
The /sys/devices/virtual handler should enter the namespaces of the
container in order to expose the correct sysfs resources. This commit
does this, by having the /sys/devices/virtual handler invoke the
passthrough handler, which in turn launches an nsenter agent in
the container's namespaces.

The nsenter agent deployed by the passthrough handler was improved
such that it will create a mount-namespace and remount procfs and
sysfs as appropriate, so that these filesystems become aware of
the resources associated with the container namespaces.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>