Fix in unmount syscall interception to enable pivot-root.
Previously, sysbox-fs, when configured with "--allow-immutable-unmounts=false",
would disallow unmounts of the container "/" since it's considered an immutable
mount.

However, this causes a problem: a program that executes pivot-root inside a
Sysbox container would fail, because the program would switch to a new root, but
then fail when unmounting the old root because that old root is the container's
"/", and Sysbox would block that operation.

In fact, this caused the Docker-in-Docker official image (docker:24-dind) to
fail when running inside a Sysbox container (again, assuming Sysbox is
configured with "--allow-immutable-unmounts=false").

This commit fixes this, by having sysbox-fs allow unmounts of the
container's "/" even if "--allow-immutable-unmounts=false".

Turns out this is safe (i.e., does not reduce security) because:

* If a process in the sysbox container tries to unmount "/", sysbox-fs will not
  veto this, but the kernel will still disallow it. The reason is that the
  process is itself running with that root, so it can't be unmounted (i.e., the
  mount is busy). Even if in the future the kernel were to change and allow the
  unmount of "/", the only thing the process would uncover is the (usually
  empty) directory at host level on which the container's rootfs was mounted.

* The only way for a process inside a Sysbox container to unmount "/" is for a
  process to pivot-root into a new root and then unmount the old root. But since
  during the pivot root the old root is mounted on a container's subdir,
  unmounting the old root only uncovers the contents of that empty subdir.

Together with this commit, integration tests have been written to verify
all what is said above.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>