Add non-seek flag to handler Open() method.
Some files that sysbox-fs emulates inside a container (e.g., /proc/uptime) are
not seekable. Add a flag to the handler's Open() method that allows the handler
to indicate if the file it's emulating is seekable or not.

This change is required to make emulation of /proc/uptime work inside
containers, starting with Linux kernel 6.5+. It's not clear what changed in that
kernel that causes this requirement, but it's clear that the /proc/uptime
handler must specify the file is non-seekable in order to the emulation to work
correctly.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>